refactor: ♻️ extract useRestaurantPage hook and fix any types in restaurant components
- Extract useRestaurantPage hook from RestaurantPage component
  - Moves all useState, useMemo, useEffect, and tRPC calls out of the
    component into a dedicated hook (hooks/useRestaurantPage.ts)
  - Relocates the getCurrentPeriod helper into the hook file with a JSDoc
  - RestaurantPage is now a pure layout/render component (~160 lines vs 326)

- Fix any[] types across restaurant sub-components
  - restaurant-controls.tsx: stations: Station[], hallEvents: Event[]
  - desktop-tabs.tsx: stations: Station[]
  - mobile-actions.tsx: dishes: DishWithRating[], stations: Station[],
    hallEvents: Event[]

- Additional fixes surfaced by stricter typing
  - dishes-view.tsx: relax activeStation to Station | undefined (can be
    undefined while data is loading)
  - mobile-actions.tsx: fix event.shortDescription → event.description
    to match the actual Event type from @peterplate/validators
  - restaurant-controls.tsx: remove misleading //...props comments;
    improve inline comments for clarity

- Remove unnecessary hallData useMemo wrapper in restaurant-page.tsx;
  replace with a plain derived value
